+++ surelog-sys/surelog_cpp_wrapper/surelog_cpp_wrapper.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cstddef>
+// #include <cstddef>
 #include <uhdm/vpi_user.h>
 #include <uhdm/sv_vpi_user.h>
 
@@ -9,7 +9,7 @@ struct SystemVerilogDesign;
 extern "C" {
 
 SystemVerilogDesign* design_create();
-vpiHandle design_compile(SystemVerilogDesign* d, char const* const* sources, size_t sources_len);
+vpiHandle design_compile(SystemVerilogDesign* d, char const* const* sources, unsigned long long sources_len);
 void design_free(SystemVerilogDesign* d);
 
 vpiHandle design_top_entity(vpiHandle design);
